scalp_edge/requirements.txt
flask>=3.0.0
PyGithub>=2.0.0
beautifulsoup4>=4.12.0
